Remove duplicated information in net databag
We store in the network databag IP address mappings with duplicated
information. When the key is an IP address we also store the same IP in
a field. When the key is the FQDN we also store the FQDN as a field.
Those duplicated fields do not seem to be used elsewhere.

Also changed the filed named "machine" to "fqdn". That name was
incorrect because we could have virtual IPs and FQDNs for them in the
"machine" field. It makes more sense just to name it FQDN.